+++ src/collective/volto/formsupport/captcha/__init__.py
@@ -1,4 +1,4 @@
-class CaptchaSupport(object):
+class CaptchaSupport:
     def __init__(self, context, request):
         self.context = context
         self.request = request