+++ src/collective/volto/formsupport/datamanager/catalog.py
@@ -1,24 +1,23 @@
-# -*- coding: utf-8 -*-
-
+from collective.volto.formsupport import logger
+from collective.volto.formsupport.interfaces import IFormDataStore
+from collective.volto.formsupport.utils import get_blocks
 from copy import deepcopy
 from datetime import datetime
-
 from plone.dexterity.interfaces import IDexterityContent
 from plone.restapi.deserializer import json_body
 from repoze.catalog.catalog import Catalog
 from repoze.catalog.indexes.field import CatalogFieldIndex
 from souper.interfaces import ICatalogFactory
-from souper.soup import NodeAttributeIndexer, Record, get_soup
+from souper.soup import get_soup
+from souper.soup import NodeAttributeIndexer
+from souper.soup import Record
 from zope.component import adapter
-from zope.interface import Interface, implementer
-
-from collective.volto.formsupport import logger
-from collective.volto.formsupport.interfaces import IFormDataStore
-from collective.volto.formsupport.utils import get_blocks
+from zope.interface import implementer
+from zope.interface import Interface
 
 
 @implementer(ICatalogFactory)
-class FormDataSoupCatalogFactory(object):
+class FormDataSoupCatalogFactory:
     def __call__(self, context):
         #  do not set any index here..maybe on each form
         catalog = Catalog()
@@ -29,7 +28,7 @@ class FormDataSoupCatalogFactory(object):
 
 @implementer(IFormDataStore)
 @adapter(IDexterityContent, Interface)
-class FormDataStore(object):
+class FormDataStore:
     def __init__(self, context, request):
         self.context = context
         self.request = request