+++ src/collective/volto/formsupport/restapi/serializer/blocks.py
@@ -1,20 +1,17 @@
-# -*- coding: utf-8 -*-
-import os
-
+from collective.volto.formsupport.interfaces import ICaptchaSupport
+from collective.volto.formsupport.interfaces import ICollectiveVoltoFormsupportLayer
 from plone import api
+from plone.base.interfaces import IPloneSiteRoot
 from plone.restapi.behaviors import IBlocks
 from plone.restapi.interfaces import IBlockFieldSerializationTransformer
-from Products.CMFPlone.interfaces import IPloneSiteRoot
-from zope.component import adapter, getMultiAdapter
+from zope.component import adapter
+from zope.component import getMultiAdapter
 from zope.interface import implementer
 
-from collective.volto.formsupport.interfaces import (
-    ICaptchaSupport,
-    ICollectiveVoltoFormsupportLayer,
-)
+import os
 
 
-class FormSerializer(object):
+class FormSerializer:
     """ """
 
     order = 200  # after standard ones